requirements.txt
setuptools
attrs